Added syntactic environments.
Environments are mappings from identifier strings to identifier objects. An
identifer object can represent an identifier bound to a value, a variable, an
inaccessible variable or an unbound indentifier.

An environment can be initialized with the bindings of another environment.
This can be done in two ways, depending on whether the new environment
represents a block or a function scope. In a block scope all bindings are
inherited and in a function scope both value and variable bindings are
inherited, but variable bindings are marked as inaccessible.